fixing typescript errors
