+++ modules/github/types.ts
@@ -1,0 +1,9 @@
+export interface GitHubFile {
+  name: string;
+  path: string;
+  sha: string;
+  size: number;
+  type: 'file' | 'dir' | 'symlink' | 'submodule';
+  content?: string;
+  url?: string;
+}
