+++ modules/playground/lib/editor-config.ts
@@ -1,4 +1,5 @@
 import type { Monaco } from "@monaco-editor/react";
+import type { editor } from 'monaco-editor'
 
 export const getEditorLanguage = (fileExtension: string): string => {
   const extension = fileExtension.toLowerCase();
@@ -262,7 +263,7 @@ export const configureMonaco = (monaco: Monaco) => {
   });
 };
 
-export const defaultEditorOptions = {
+export const defaultEditorOptions : editor.IStandaloneEditorConstructionOptions = {
   // Font settings
   fontSize: 14,
   fontFamily: "'JetBrains Mono', 'Fira Code', 'SF Mono', Consolas, 'Liberation Mono', Menlo, Courier, monospace",
@@ -280,7 +281,7 @@ export const defaultEditorOptions = {
   padding: { top: 16, bottom: 16 },
   
   // Line settings
-  lineNumbers: "on",
+  lineNumbers: "on" ,
   lineHeight: 20,
   renderLineHighlight: "all",
   renderWhitespace: "selection",
@@ -309,13 +310,13 @@ export const defaultEditorOptions = {
   // Selection
   multiCursorModifier: "ctrlCmd",
   selectionHighlight: true,
-  occurrencesHighlight: true,
+  occurrencesHighlight: "singleFile",
   
   // Suggestions
   suggestOnTriggerCharacters: true,
   acceptSuggestionOnEnter: "on",
   tabCompletion: "on",
-  wordBasedSuggestions: true,
+  wordBasedSuggestions: "currentDocument",
   quickSuggestions: {
     other: true,
     comments: false,
@@ -333,8 +334,8 @@ export const defaultEditorOptions = {
   },
   
   // Guides
-  renderIndentGuides: true,
-  highlightActiveIndentGuide: true,
+  
+  
   rulers: [80, 120],
   
   // Performance
@@ -346,7 +347,7 @@ export const defaultEditorOptions = {
   
   // Cursor
   cursorBlinking: "smooth",
-  cursorSmoothCaretAnimation: true,
+  cursorSmoothCaretAnimation: "on",
   cursorStyle: "line",
   cursorWidth: 2,
