+++ modules/webContainers/hooks/transformer.ts
@@ -1,59 +1,40 @@
-interface TemplateItem {
-  filename: string;
-  fileExtension: string;
-  content: string;
-  folderName?: string;
-  items?: TemplateItem[];
-}
+import type { TemplateFile, TemplateFolder, TemplateItem } from "@/modules/playground/lib/path-to-json";
 
 interface WebContainerFile {
-  file: {
-    contents: string;
-  };
+  file: { contents: string };
 }
 
 interface WebContainerDirectory {
-  directory: {
-    [key: string]: WebContainerFile | WebContainerDirectory;
-  };
+  directory: { [key: string]: WebContainerFile | WebContainerDirectory };
 }
 
 type WebContainerFileSystem = Record<string, WebContainerFile | WebContainerDirectory>;
 
-export function transformToWebContainerFormat(template: { folderName: string; items: TemplateItem[] }): WebContainerFileSystem {
+export function transformToWebContainerFormat(template: TemplateFolder): WebContainerFileSystem {
   function processItem(item: TemplateItem): WebContainerFile | WebContainerDirectory {
-    if (item.folderName && item.items) {
-      // This is a directory
+    if ("folderName" in item) {
+      // TypeScript now knows item is TemplateFolder
       const directoryContents: WebContainerFileSystem = {};
-      
       item.items.forEach(subItem => {
-        const key = subItem.fileExtension 
-          ? `${subItem.filename}.${subItem.fileExtension}`
-          : subItem.folderName!;
+        const key = "folderName" in subItem
+          ? subItem.folderName
+          : `${subItem.filename}.${subItem.fileExtension}`;
         directoryContents[key] = processItem(subItem);
       });
-
-      return {
-        directory: directoryContents
-      };
+      return { directory: directoryContents };
     } else {
-      // This is a file
-      return {
-        file: {
-          contents: item.content
-        }
-      };
+      // TypeScript now knows item is TemplateFile
+      return { file: { contents: item.content } };
     }
   }
 
   const result: WebContainerFileSystem = {};
-  
   template.items.forEach(item => {
-    const key = item.fileExtension 
-      ? `${item.filename}.${item.fileExtension}`
-      : item.folderName!;
+    const key = "folderName" in item
+      ? item.folderName
+      : `${item.filename}.${item.fileExtension}`;
     result[key] = processItem(item);
   });
 
   return result;
-}1
+}
